Improvement

+++ Basic/Options.cs
@@ -7,9 +7,9 @@
         public readonly double Alfa;
         public readonly double Beta;
         public readonly double Ro;
-        public readonly decimal Delta;
+        public readonly double Delta;
 
-        public Options(int numberOfIterations, int numberOfRegions, double alfa, double beta, double ro, decimal delta)
+        public Options(int numberOfIterations, int numberOfRegions, double alfa, double beta, double ro, double delta)
         {
             NumberOfIterations = numberOfIterations;
             NumberOfRegions = numberOfRegions;
